requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2026.3
cryptography==46.0.7
daphne==4.2.1
dj-database-url==3.1.2
Django
django-allauth==65.15.1
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-debug-toolbar==6.3.0
django-environ==0.13.0
django-phonenumber-field==8.4.0
django-ratelimit==4.1.0
djangorestframework==3.17.1
frozenlist==1.8.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
multidict==6.7.1
packaging==26.0
phonenumbers==9.0.27
pillow==12.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
twilio==9.10.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
yarl==1.23.0
zope.interface==8.3
Pillow
gunicorn
whitenoise
psycopg2-binary
python-dotenv
dj-database-url
django-extensions
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
groq